Change sample-server/sample-client message lengths to uint16_t.
This commit moves the client and server message-lengths from uint32_t
to uint16_t. This matches changes made in previous commits to the main
library.

+++ test/socks_client.c
@@ -15,7 +15,7 @@ int main(int argc, char **argv)
     ssize_t result;
     char buffer[1024];
     char *cmd;
-    uint32_t cmd_len;
+    uint16_t cmd_len;
 
     progname = basename(argv[0]);
